requirements.txt
django
django-debug-toolbar
django-filter
django-otp
django-templated-mail
djangorestframework
djangorestframework-simplejwt
djoser
drf-nested-routers
gunicorn
Markdown
mysqlclient
oauthlib
packaging
pycparser
PyJWT
PyMySQL
python-dotenv
celery
redis
flower